@@ -9,4 +9,4 @@ __author_email__ = "charlie.clark@clark-consulting.eu"
 __license__ = "MIT/Expat"
 __maintainer_email__ = "openpyxl-users@googlegroups.com"
 __url__ = "https://openpyxl.readthedocs.io"
-__version__ = "2.6.3"
+__version__ = "3.0.0dev"
